refactor: inline kernel identity into approved native source
